refactor(push): move push service wiring to the composition root
push.service.ts no longer imports infrastructure (push-sender, logger,
queue). createPushService now requires its deps, and the production
singleton is constructed in domain/services/index.ts and exported from the
barrel. Consumers (push routes, evening-nudge worker, entrypoint) import
pushService from domain/services/index.js, matching how every other domain
service is consumed.
