+++ packages/backend/src/infrastructure/queue/workers/evening-nudge-logic.ts
@@ -1,6 +1,6 @@
 import { db } from '../../database/connection.js'
 import { queueLogger } from '../../logger/logger.js'
-import { pushService } from '../../../domain/services/push.service.js'
+import { pushService } from '../../../domain/services/index.js'
 import { buildEveningNudge } from '../../../domain/services/evening-nudge-copy.js'
 import { loadFeaturedLeader } from './featured-leader.js'
